Update Fruit database

+++ src/main/kotlin/ktorwebapi/models/Fruit.kt
@@ -1,3 +1,3 @@
 package ktorwebapi.models
 
-data class Fruit(val id: Long, val no: String, val description: String)
+data class Fruit(val id: Int, val no: String, val description: String)
